Permission added

+++ app/Http/Controllers/PermissionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Permission;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -14,17 +15,20 @@ class PermissionController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         $permissions = Permission::latest()->get();
         return view('permission.index', compact(['permissions']));
     }
 
     public function create()
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         return view('permission.create');
     }
 
     public function store(Request $request)
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         $permission = new Permission();
         $permission->title = $request->title;
         $permission->status = $request->status;
@@ -36,19 +40,21 @@ class PermissionController extends Controller
 
     }
 
-    public function show()
-    {
+    // public function show()
+    // {
 
-    }
+    // }
 
     public function edit($id)
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         $permission = Permission::find($id);
         return view('permission.edit', compact(['permission']));
     }
 
     public function update(Request $request, $id)
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         $permission = Permission::find($request->id);
         $permission->title = $request->title;
         $permission->status = $request->status;
@@ -61,6 +67,7 @@ class PermissionController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('permission-access'), redirect('error'));
         $permission = Permission::find($id);
         $permission->delete();
